backend/requirements.txt
APIFlask==2.2.1
apispec==6.7.1
blinker==1.9.0
click==8.1.7
Flask==3.1.0
Flask-HTTPAuth==4.8.0
flask-marshmallow==1.2.1
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==3.0.2
marshmallow==3.23.1
numpy==2.1.3
packaging==24.2
python-dotenv==1.0.1
scipy==1.14.1
threadpoolctl==3.5.0
webargs==8.6.0
Werkzeug==3.1.3
